requirements.txt
anyio==4.3.0
asgiref==3.5.2
CacheControl==0.14.3
cachetools==5.3.3
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.1.0
chatbot_session_flow==0.1.6
cloudinary==1.30.0
confusable-homoglyphs==3.2.0
cryptography==37.0.4
cssselect==1.2.0
cssutils==2.11.0
dj-database-url==0.5.0
Django==4.1.2
django-appconf==1.0.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-imagekit==5.0.0
django-on-heroku==1.1.2
django-registration==3.3
django-simple-mail==2.3.5
django-tinymce==3.4.0
djangorestframework==3.14.0
djangorestframework-pagination==1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==20.1.0
h11==0.14.0
html2text==2024.2.26
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
lxml==5.2.2
msgpack==1.1.0
pilkit==3.0
pillow==10.3.0
premailer==3.10.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.15
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.5
pytz==2022.6
pyuploadcare==3.1.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
